+++ src/test/kotlin/com/github/jensim/megamanipulator/actions/vcs/githubcom/GithubComClientTest.kt
@@ -90,6 +90,8 @@ class GithubComClientTest {
                 limit = 10,
                 state = CodeHostSettingsType.GITHUB.prStateOpen,
                 role = githubSettings.codeHostType.prRoleAuthor,
+                project = null,
+                repo = null,
             )
         }