+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 # mega-manipulator Changelog
 
 ## [Unreleased]
+### Added
+- Configurable http concurrency for GitHub Code host to tackle slow loading of PRs and also the over-concurrent http requests for fetching forks
+- Fetch pull requests from a specific project or repo (GitHub & GitLab)
 
 ## [0.0.53]
 ### Improvement